adding documentation in the landing page, also moving orutes to main.py, moving schema to transform folder
