+++ requirements.txt
@@ -1,4 +1,4 @@
-fastapi==0.112.2
+fastapi[standard]==0.112.2
 selenium==4.23.1
 beautifulsoup4==4.12.3
 pydantic==2.8.2
