--- src/api/__init__.py
@@ -1,1 +1,0 @@
-from .routes import app as api_router
