+++ src/docs/static/styles.css
@@ -1,0 +1,28 @@
+body {
+    font-family: Arial, sans-serif;
+    background-color: #f4f4f4;
+    margin: 0;
+    padding: 20px;
+}
+.bash{
+    display: inline-block;
+    padding: 10px;
+    padding-left : 30px ;
+    padding-right: 30px  ;
+    font-size: 20px;
+    background-color: gainsboro;
+}
+hr {
+    border: none;
+    height: 1px; 
+    background-color: #000; 
+    margin: 50px 0; 
+}
+a{
+    text-decoration: none;
+    color: black;
+}
+a:hover{
+    text-decoration-line: underline;
+    color:blue;
+}
