--- src/service/transform/formats.py
@@ -1,26 +1,0 @@
-def get_data_pool_format():
-    return {
-        "tracking_url": None,
-        "text": [], 
-        "img" :[]
-    }
-
-
-def get_data_bucket_format():
-    return {
-       "product_url" : None,
-        "product_name": None,
-        "seller_name" : None,
-        "is_ad": False,
-        "price": None,
-        "price_int" : None,
-        "discount": None,
-        "discount_float" : None,
-        "sold_items": None,
-        "rating": None,
-        "location": None,
-        "seller_url": None,
-        "image_url" : None
-    }
-
-
