+++ src/service/transform/transform_html.py
@@ -1,7 +1,8 @@
 from .import process_raw_data
 from . import raw_data_pool
 from bs4 import BeautifulSoup
-
+from .schema import ProductData
+from . import schema
 
 def tokopedia(html):
     data_bucket_list = []
@@ -15,6 +16,7 @@ def tokopedia(html):
     for product_div in all_product_div:
         data_bucket = raw_data_pool.collect_data(product_div)
         processed_data_bucket = process_raw_data.get_processed_data(data_bucket)
-        data_bucket_list.append(processed_data_bucket)
-
+        if ProductData.is_valid(processed_data_bucket):
+            data_bucket_list.append(schema.search_product_data_schema(processed_data_bucket))
+            
     return data_bucket_list
